refactor: improve toolings
