--- .build.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "extends": [
-    "./.eslintrc.json"
-  ],
-  "rules": {
-    "no-debugger": "error"
-  }
-}
