+++ .editorconfig
@@ -1,19 +1,16 @@
 # EditorConfig helps developers define and maintain consistent
 # coding styles between different editors and IDEs
-# http://editorconfig.org
+# https://editorconfig.org
 
 root = true
 
 [*]
 # Change these settings to your own preference
-indent_style = space
 indent_size = 2
+indent_style = space
 
 # We recommend you to keep these unchanged
-end_of_line = lf
 charset = utf-8
-trim_trailing_whitespace = true
+end_of_line = lf
 insert_final_newline = true
-
-[*.md]
-trim_trailing_whitespace = false
+trim_trailing_whitespace = true
