+++ .eslintrc.json
@@ -1,59 +1,5 @@
 {
-  "env": {
-    "browser": true
-  },
   "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:promise/recommended",
-    "plugin:regexp/recommended"
-  ],
-  "ignorePatterns": [
-    "**/*.mjs",
-    "vite.config.ts"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "./tsconfig.json"
-  },
-  "plugins": [
-    "@typescript-eslint",
-    "simple-import-sort"
-  ],
-  "rules": {
-    "no-async-promise-executor": "off",
-    "no-debugger": "off",
-    "no-unused-vars": "off",
-    "@typescript-eslint/no-unused-vars": [
-      "error", {
-        "argsIgnorePattern": "^_",
-        "varsIgnorePattern": "^_"
-      }
-    ],
-    "comma-dangle": "off",
-    "@typescript-eslint/comma-dangle": [
-      "error", {
-        "arrays": "always-multiline",
-        "objects": "always-multiline",
-        "imports": "always-multiline",
-        "exports": "always-multiline",
-        "functions": "never"
-      }
-    ],
-    "comma-spacing": "off",
-    "@typescript-eslint/comma-spacing": [
-      "error", {
-        "before": false,
-        "after": true
-      }
-    ],
-    "@typescript-eslint/consistent-type-imports": [
-      "error", {
-        "prefer": "type-imports",
-        "disallowTypeAnnotations": false
-      }
-    ],
-    "simple-import-sort/imports": "error",
-    "simple-import-sort/exports": "error"
-  }
+    "./node_modules/@reallyland/tools/.eslintrc.json"
+  ]
 }
