+++ .github/FUNDING.yml
@@ -1,0 +1,5 @@
+# My Github Sponsors profile is live at 2022-04-03 14:33 SGT! 🚀 🎉
+github: [motss]
+
+# Buy me a matcha to support my passion!
+custom: ['https://www.buymeacoffee.com/RLmMhgXFb']
