+++ .github/workflows/ci.yml
@@ -12,67 +12,52 @@ on:
     tags-ignore:
       - '*'
 
+concurrency:
+      group: ${{ github.workflow }}-${{ github.ref || github.run_id }}
+      cancel-in-progress: true
+
 defaults:
   run:
     shell: bash
 
 env:
-  CI_SKIP: false
-  GITHUB_COMMIT_MESSAGE: ''
   IS_UBUNTU: false
   NODE_VERSION: v0.0.0
   NPM_CACHE_DIR: ~/.npm
   NPM_VERSION: 0.0.0
 
 jobs:
-  test:
-    if: "!contains(github.event.head_commit.message, '[ci skip]') && !contains(github.event.head_commit.message, '[skip ci]') && !contains(github.event.head_commit.message, 'doc:')"
+  lint:
+    if: "!contains(github.event.head_commit.message, '[ci skip]') && !contains(github.event.head_commit.message, '[skip ci]') && !contains(github.event.head_commit.message, 'docs:') && !contains(github.event.head_commit.message, 'chore(release)')"
     env:
-      GITHUB_COMMIT_MESSAGE: ${{ github.event.head_commit.message }}
-    name: ${{ matrix.os }} ${{ matrix.target }}
+      IS_UBUNTU: ${{ contains(matrix.os, 'ubuntu') }}
+    name: lint - ${{ matrix.os }} ${{ matrix.target }}
     runs-on: ${{ matrix.os }}
     timeout-minutes: 120
     strategy:
       matrix:
-        os: [macos-latest, ubuntu-latest, windows-latest]
-        target: [14.x, 16.x]
+        os: [ubuntu-latest]
+        target: [14.x, 16.x, 18.x]
 
     steps:
-      - name: CI skip check
-        run: |
-          if [[ -n "$(echo $GITHUB_COMMIT_MESSAGE | grep -iE '^[0-9]+[.]{1,}[0-9]+[.]{1,}[0-9]+')" ]]; then
-            echo "[INFO] SKIP CI"
-            echo "CI_SKIP=true" >> $GITHUB_ENV
-          fi
-
       - name: Checkout
-        if: env.CI_SKIP == 'false'
-        uses: actions/checkout@v2
-
-      - name: Setup node@${{ matrix.target }}
-        if: env.CI_SKIP == 'false'
-        uses: actions/setup-node@v2
-        with:
-          node-version: ${{ matrix.target }}
-
-      # - name: Install latest npm
-      #   if: env.CI_SKIP == 'false'
-      #   run: |
-      #     npm i -g npm@latest
+        uses: actions/checkout@v3
 
       - name: List versions
-        if: env.CI_SKIP == 'false'
         run: |
-          pwd && ls -la
-          echo "IS_UBUNTU=${{ contains(matrix.os, 'ubuntu') }}" >> $GITHUB_ENV
           echo "NODE_VERSION=$(node -v)" >> $GITHUB_ENV
           echo "NPM_CACHE_DIR=$(npm config get cache)" >> $GITHUB_ENV
           echo "NPM_VERSION=$(npm -v)" >> $GITHUB_ENV
 
+          echo "${{ github.workflow }}-${{ github.ref || github.run_id }}"
+          pwd
+          ls -la
+          which npm
+          # npm x -y -- envinfo@latest
+
       - name: Cache dependencies
-        if: env.CI_SKIP == 'false'
         id: npm-cache
-        uses: actions/cache@v2
+        uses: actions/cache@v3
         with:
           # See this glob workaround at https://github.com/actions/toolkit/issues/713.
           path: |
@@ -83,31 +68,127 @@ jobs:
             ${{ matrix.os }}-node-${{ env.NODE_VERSION }}-npm-${{ env.NPM_VERSION }}
 
       - name: Install dependencies
-        if: env.CI_SKIP == 'false'
         run: |
-          bash ./list-npx-cache.sh
           npm ci
-          bash ./list-npx-cache.sh
+          sh $(npm root)/@reallyland/tools/list-npx-cache.sh
 
       - name: Lint
-        if: env.CI_SKIP == 'false'
         run: |
           npm run lint:build
 
-      - name: Test
-        if: env.CI_SKIP == 'false' && env.IS_UBUNTU == 'false'
+  test:
+    if: "!contains(github.event.head_commit.message, '[ci skip]') && !contains(github.event.head_commit.message, '[skip ci]') && !contains(github.event.head_commit.message, 'docs:') && !contains(github.event.head_commit.message, 'chore(release)')"
+    env:
+      IS_UBUNTU: ${{ contains(matrix.os, 'ubuntu') }}
+    name: test - ${{ matrix.os }} ${{ matrix.target }}
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        target: [14.x, 16.x, 18.x]
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      # - name: Setup node@${{ matrix.target }}
+      #   uses: actions/setup-node@v3
+      #   with:
+      #     node-version: ${{ matrix.target }}
+
+      - name: List versions
         run: |
-          npm t
+          echo "NODE_VERSION=$(node -v)" >> $GITHUB_ENV
+          echo "NPM_CACHE_DIR=$(npm config get cache)" >> $GITHUB_ENV
+          echo "NPM_VERSION=$(npm -v)" >> $GITHUB_ENV
+
+          echo "${{ github.workflow }}-${{ github.ref || github.run_id }}"
+          pwd
+          ls -la
+          which npm
+          # npm x -y -- envinfo@latest
+
+      - name: Cache dependencies
+        id: npm-cache
+        uses: actions/cache@v3
+        with:
+          # See this glob workaround at https://github.com/actions/toolkit/issues/713.
+          path: |
+            ${{ env.NPM_CACHE_DIR }}/*
+            !${{ env.NPM_CACHE_DIR }}/_npx
+          key: ${{ matrix.os }}-node-${{ env.NODE_VERSION }}-npm-${{ env.NPM_VERSION }}-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ matrix.os }}-node-${{ env.NODE_VERSION }}-npm-${{ env.NPM_VERSION }}
+
+      - name: Install dependencies
+        run: |
+          npm ci
+          sh $(npm root)/@reallyland/tools/list-npx-cache.sh
+
+      - name: Test
+        if: env.IS_UBUNTU == 'false'
+        run: |
+          npm t -- --no-coverage --no-watch
 
       - name: Test with coverage
-        if: env.CI_SKIP == 'false' && env.IS_UBUNTU != 'false'
+        if: env.IS_UBUNTU != 'false'
         run: |
-          npm run test:coverage
+          npm t -- --color --no-watch
 
       - name: Upload coverage to codecov
-        if: env.CI_SKIP == 'false' && env.IS_UBUNTU != 'false' && success()
-        uses: codecov/codecov-action@v1
+        if: env.IS_UBUNTU != 'false' && success()
+        uses: codecov/codecov-action@v3
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
           file: ./coverage/lcov.info
           flags: unit_tests
+
+  bench:
+    if: "!contains(github.event.head_commit.message, '[ci skip]') && !contains(github.event.head_commit.message, '[skip ci]') && !contains(github.event.head_commit.message, 'docs:') && !contains(github.event.head_commit.message, 'chore(release)')"
+    env:
+      IS_UBUNTU: ${{ contains(matrix.os, 'ubuntu') }}
+    name: bench - ${{ matrix.os }} ${{ matrix.target }}
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        target: [14.x, 16.x, 18.x]
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: List versions
+        run: |
+          echo "NODE_VERSION=$(node -v)" >> $GITHUB_ENV
+          echo "NPM_CACHE_DIR=$(npm config get cache)" >> $GITHUB_ENV
+          echo "NPM_VERSION=$(npm -v)" >> $GITHUB_ENV
+
+          echo "${{ github.workflow }}-${{ github.ref || github.run_id }}"
+          pwd
+          ls -la
+          which npm
+          # npm x -y -- envinfo@latest
+
+      - name: Cache dependencies
+        id: npm-cache
+        uses: actions/cache@v3
+        with:
+          # See this glob workaround at https://github.com/actions/toolkit/issues/713.
+          path: |
+            ${{ env.NPM_CACHE_DIR }}/*
+            !${{ env.NPM_CACHE_DIR }}/_npx
+          key: ${{ matrix.os }}-node-${{ env.NODE_VERSION }}-npm-${{ env.NPM_VERSION }}-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ matrix.os }}-node-${{ env.NODE_VERSION }}-npm-${{ env.NPM_VERSION }}
+
+      - name: Install dependencies
+        run: |
+          npm ci
+          sh $(npm root)/@reallyland/tools/list-npx-cache.sh
+
+      - name: Benchmark
+        run: |
+          npm run bench -- --no-watch
