+++ .gitignore
@@ -1,19 +1,23 @@
 # .gitignore 101 - https://www.quora.com/Does-gitignore-file-only-work-on-current-folder
 
-# dot files
+# dot files/ folders
 .*cache
 .build/
 .DS_Store
+.firebase/
+.nyc_output
 .tmp/
-.vscode
 *.env
-
-# debug
-*-debug.log*
+*.idea
+*.local
+*.log*
+*.tsbuildinfo
 
 # folders
 coverage*/
 dist*/
+log*/
 node_modules/
+web_modules/
 
 # Do not ignore
