+++ .npmrc
@@ -1,6 +1,9 @@
 # User config
 foreground-scripts=true
+message="chore(release): %s :tada:"
+package-lock=true
 preid=rc
 progress=true
 quiet=true
+sign-git-tag=true
 update-binary=true
