--- .swcrc
@@ -1,15 +1,0 @@
-{
-  "jsc": {
-    "parser": {
-      "syntax": "typescript",
-      "tsx": true,
-      "decorators": true,
-      "dynamicImport": true
-    },
-    "target": "es2019",
-    "externalHelpers": false
-  },
-  "module": {
-    "type": "commonjs"
-  }
-}
