+++ package.json
@@ -15,6 +15,7 @@
     "type": "git",
     "url": "git@github.com:motss/normalize-diacritics.git"
   },
+  "funding": "https://github.com/motss/lit-ntml?sponsor=1",
   "license": "MIT",
   "author": {
     "name": "Rong Sen Ng (motss)",
@@ -25,8 +26,24 @@
   "type": "module",
   "exports": {
     ".": {
+      "types": "./dist/index.d.ts",
       "import": "./dist/index.js",
       "default": "./dist/index.js"
+    },
+    "./diacritics": {
+      "types": "./dist/diacritics.d.ts",
+      "import": "./dist/diacritics.js",
+      "default": "./dist/diacritics.js"
+    },
+    "./normalize": {
+      "types": "./dist/normalize.d.ts",
+      "import": "./dist/normalize.js",
+      "default": "./dist/normalize.js"
+    },
+    "./normalizeSync": {
+      "types": "./dist/normalizeSync.d.ts",
+      "import": "./dist/normalizeSync.js",
+      "default": "./dist/normalizeSync.js"
     },
     "./dist/*": "./dist/*"
   },
@@ -38,53 +55,31 @@
     "dist/*.*js",
     "dist/*.d.ts.map",
     "dist/*.d.ts",
-    "postinstall.*",
+    "!**/*bench*/**/*.*",
     "!**/*test*/**/*.*"
   ],
   "scripts": {
+    "bench": "vitest bench",
     "build": "npm run clean && tsc",
-    "clean": "rm -rf .*cache *.log .swc/ coverage/ dist/ logs/",
-    "postinstall": "sh postinstall.sh",
-    "lint": "eslint src --ext .js,.ts",
-    "lint:build": "npm run lint -- --config .build.eslintrc.json",
-    "lint:commit": "npm x -y -- commitlint@latest --edit",
-    "pre-commit": "package-check && nano-staged && tsc --noEmit",
-    "prepublishOnly": "package-check && npm run lint:build && npm run build",
-    "test": "uvu -r @swc/register src/tests",
-    "test:coverage": "c8 --reporter=lcov --exclude=src/tests npm t",
-    "test:watch": "npm run clean && watchlist src -- npm t",
-    "watch": "tsc --watch"
-  },
-  "simple-git-hooks": {
-    "pre-commit": "npm run pre-commit",
-    "commit-msg": "npm run lint:commit"
-  },
-  "commitlint": {
-    "extends": [
-      "@commitlint/config-conventional"
-    ]
+    "clean": "sh $(npm root)/@reallyland/tools/cleanup.sh",
+    "postinstall": "FILE=\"$(npm root)/@reallyland/tools/postinstall.sh\"; if [ -f \"$FILE\" ]; then sh \"$FILE\"; fi",
+    "lint": "sh $(npm root)/@reallyland/tools/lint-build.sh -c $(npm root)/@reallyland/tools/.eslintrc.json -s 'src/**/*.ts'",
+    "lint-commit": "sh $(npm root)/@reallyland/tools/lint-commit.sh",
+    "lint:build": "sh $(npm root)/@reallyland/tools/lint-build.sh -c $(npm root)/@reallyland/tools/browser/.build.eslintrc.json",
+    "pre-commit": "sh $(npm root)/@reallyland/tools/pre-commit.sh",
+    "prepublishOnly": "npm run lint:build && npm run build",
+    "test": "vitest --coverage",
+    "ts": "tsc",
+    "version": "sh $(npm root)/@reallyland/tools/generate-changelogs.sh && git add *CHANGELOG.md"
   },
   "dependencies": {
-    "tslib": "^2.1.0"
+    "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@commitlint/config-conventional": "^16.2.1",
-    "@reallyland/tsconfig": "^4.0.1",
-    "@skypack/package-check": "^0.2.2",
-    "@swc/core": "^1.2.148",
-    "@swc/register": "^0.1.10",
-    "@types/node": "^17.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
-    "c8": "^7.11.0",
-    "eslint": "^8.10.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-regexp": "^1.5.1",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
-    "nano-staged": "^0.6.0",
-    "typescript": "^4.6.2",
-    "uvu": "^0.5.3",
-    "watchlist": "^0.3.0"
+    "@reallyland/tools": "^0.1.1",
+    "@vitest/coverage-c8": "^0.24.3",
+    "typescript": "^4.8.4",
+    "vitest": "^0.24.3"
   },
   "engines": {
     "node": ">= 14.x",
@@ -94,6 +89,9 @@
     "access": "public"
   },
   "nano-staged": {
-    "*.ts": "npm run lint:build -- --fix"
+    "*.ts": [
+      "npm run lint:build -- -f 1",
+      "git add"
+    ]
   }
 }
