--- postinstall.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-npx -y -- zx@latest ./postinstall.mjs
