+++ src/normalize.ts
@@ -1,0 +1,5 @@
+import { normalizeSync } from './normalize-sync.js';
+
+export async function normalize(input?: string | null): Promise<string> {
+  return normalizeSync(input);
+}
