+++ src/tests/types.ts
@@ -1,0 +1,4 @@
+export interface TestCase<In, Out> {
+  input: In;
+  output: Out;
+}
