+++ src/types.ts
@@ -1,0 +1,4 @@
+export interface Diacritics {
+  diacritics: RegExp;
+  letter: string;
+}
