+++ vite.config.ts
@@ -1,0 +1,11 @@
+/// <reference types="vitest" />
+
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  test: {
+    coverage: {
+      reporter: ['lcov', 'text'],
+    },
+  },
+})
